pyproject.toml
[project]
name = "gigs-bot"
version = "0.1.0"
description = "Telegram bot for Google Calendar management with Gemini AI"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.13.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "redis[hiredis]>=5.0.0",
    "pydantic-settings>=2.0.0",
    "google-auth>=2.25.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.110.0",
    "google-genai>=1.0.0",
    "cryptography>=42.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
